requirements.txt
PyGithub>=2.1.1
rich>=13.7.0
pyyaml>=6.0.1
pytest>=7.4.0 